Undid some automatic formatting of code from my IDE
